// app/src/generated/services/Ben_ltkissuefilesService.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

import type { Ben_ltkissuefilesBase, Ben_ltkissuefiles } from '../models/Ben_ltkissuefilesModel';
import type { Ben_ltkissuefilesUploadColumnName } from '../models/Ben_ltkissuefilesModel';
import type { GetEntityMetadataOptions, EntityMetadata } from '@microsoft/power-apps/data/metadata/dataverse';
import type { IGetOptions, IGetAllOptions } from '../models/CommonModels';
import type { IOperationResult } from '@microsoft/power-apps/data';
import { dataSourcesInfo } from '../../../.power/schemas/appschemas/dataSourcesInfo';
import { getClient } from '@microsoft/power-apps/data';


export class Ben_ltkissuefilesService {
  private static readonly dataSourceName = 'ben_ltkissuefiles';

  private static readonly client = getClient(dataSourcesInfo);

  public static async create(record: Omit<Ben_ltkissuefilesBase, 'ben_ltkissuefileid'>): Promise<IOperationResult<Ben_ltkissuefiles>> {
    const result = await Ben_ltkissuefilesService.client.createRecordAsync<Omit<Ben_ltkissuefilesBase, 'ben_ltkissuefileid'>, Ben_ltkissuefiles>(
      Ben_ltkissuefilesService.dataSourceName,
      record
    );
    return result;
  }

  public static async update(id: string, changedFields: Partial<Omit<Ben_ltkissuefilesBase, 'ben_ltkissuefileid'>>): Promise<IOperationResult<Ben_ltkissuefiles>> {
    const result = await Ben_ltkissuefilesService.client.updateRecordAsync<Partial<Omit<Ben_ltkissuefilesBase, 'ben_ltkissuefileid'>>, Ben_ltkissuefiles>(
      Ben_ltkissuefilesService.dataSourceName,
      id,
      changedFields
    );
    return result;
  }

  public static async delete(id: string): Promise<void> {
    await Ben_ltkissuefilesService.client.deleteRecordAsync(
      Ben_ltkissuefilesService.dataSourceName,
      id);
  }

  public static async get(id: string, options?: IGetOptions): Promise<IOperationResult<Ben_ltkissuefiles>> {
    const result = await Ben_ltkissuefilesService.client.retrieveRecordAsync<Ben_ltkissuefiles>(
      Ben_ltkissuefilesService.dataSourceName,
      id,
      options
    );
    return result;
  }

  public static async getAll(options?: IGetAllOptions): Promise<IOperationResult<Ben_ltkissuefiles[]>> {
    const result = await Ben_ltkissuefilesService.client.retrieveMultipleRecordsAsync<Ben_ltkissuefiles>(
      Ben_ltkissuefilesService.dataSourceName,
      options
    );
    return result;
  }

  public static getMetadata(
    options: GetEntityMetadataOptions<Ben_ltkissuefiles> = {}
  ): Promise<IOperationResult<Partial<EntityMetadata>>> {
    return Ben_ltkissuefilesService.client.executeAsync({
      dataverseRequest: {
        action: "getEntityMetadata",
        parameters: {
          tableName: Ben_ltkissuefilesService.dataSourceName,
          options: options as GetEntityMetadataOptions,
        },
      },
    });
  }

  public static async upload(id: string, columnName: Ben_ltkissuefilesUploadColumnName, file: File, fileDisplayName?: string): Promise<IOperationResult<void>> {
    const arrayBuffer = await file.arrayBuffer();
    const data = new Uint8Array(arrayBuffer);
    const result = await Ben_ltkissuefilesService.client.uploadFileToRecord(
      Ben_ltkissuefilesService.dataSourceName,
      id,
      columnName,
      fileDisplayName || file.name,
      data,
    );
    return result;
  }
}